+++ recipes/RecipeCstVisitor.ts
@@ -6,8 +6,9 @@ import {lex} from "./lexer/RecipeLexer";
 /**
  * The result of the lexing, parsing, and visiting.
  */
-export type RecipeResult = {
-    recipe: Recipe | Array<Ingredient> | Array<Step>,
+export type RecipeResult<R extends Recipe | Array<Ingredient> | Array<Step>> = {
+    // recipe: Recipe | Array<Ingredient> | Array<Step>,
+    result: R,
     errors: Array<ILexingError>
 }
 
@@ -295,7 +296,7 @@ export const defaultOptions: Options = {
  * recipe or recipe fragment.
  * @return A recipe result holding the recipe object and any parsing errors
  */
-export function toRecipe(text: string, options: Options = defaultOptions): RecipeResult {
+export function convertText(text: string, options: Options = defaultOptions): RecipeResult<Recipe | Array<Ingredient> | Array<Step>> {
     const {
         deDupSections = false,
         logWarnings = false,
@@ -323,10 +324,72 @@ export function toRecipe(text: string, options: Options = defaultOptions): Recip
 
     const {cst, lexingResult} = parse(text)
 
-    const result = toAstVisitorInstance.visit(cst)
-
     return {
-        recipe: toAstVisitorInstance.visit(cst),
+        result: toAstVisitorInstance.visit(cst),
         errors: lexingResult.errors
     }
+}
+
+/**
+ * Converts the text into a {@link Recipe} that holds an array of {@link Ingredient} and
+ * an array of {@link Step}.
+ * @param text The text to convert into a recipe object
+ * @param deDupSections When set to `true` only sets the section of the first ingredient
+ * of each section to current section.
+ * @param logWarnings When set to `true` then logs warning to the console, otherwise
+ * does not log warnings. Warning and errors are reported in the returned object
+ * in either case.
+ * @return A {@link RecipeResult<Recipe>} holding the parsed text
+ * @see toIngredients
+ * @see toSteps
+ */
+export function toRecipe(
+    text: string,
+    deDupSections: boolean = false,
+    logWarnings: boolean = false
+): RecipeResult<Recipe> {
+    const options: Options = {deDupSections, logWarnings, inputType: ParseType.RECIPE}
+    return convertText(text, options) as RecipeResult<Recipe>
+}
+
+/**
+ * Converts the text into an array of {@link Ingredient}
+ * @param text The text to convert into an array of ingredients
+ * @param deDupSections When set to `true` only sets the section of the first ingredient
+ * of each section to current section.
+ * @param logWarnings When set to `true` then logs warning to the console, otherwise
+ * does not log warnings. Warning and errors are reported in the returned object
+ * in either case.
+ * @return A {@link RecipeResult<Recipe>} holding the parsed text
+ * @see toRecipe
+ * @see toSteps
+ */
+export function toIngredients(
+    text: string,
+    deDupSections: boolean = false,
+    logWarnings: boolean = false
+): RecipeResult<Array<Ingredient>> {
+    const options: Options = {deDupSections, logWarnings, inputType: ParseType.INGREDIENTS}
+    return convertText(text, options) as RecipeResult<Array<Ingredient>>
+}
+
+/**
+ * Converts the text into an array of {@link Step}.
+ * @param text The text to convert into an array of steps
+ * @param deDupSections When set to `true` only sets the section of the first ingredient
+ * of each section to current section.
+ * @param logWarnings When set to `true` then logs warning to the console, otherwise
+ * does not log warnings. Warning and errors are reported in the returned object
+ * in either case.
+ * @return A {@link RecipeResult<Recipe>} holding the parsed text
+ * @see toRecipe
+ * @see toIngredients
+ */
+export function toSteps(
+    text: string,
+    deDupSections: boolean = false,
+    logWarnings: boolean = false
+): RecipeResult<Array<Step>> {
+    const options: Options = {deDupSections, logWarnings, inputType: ParseType.STEPS}
+    return convertText(text, options) as RecipeResult<Array<Step>>
 }